move shared code to core and nits
