+++ sdk/extensions/Microsoft.Extensions.Azure/src/AzureEventSourceLogForwarder.cs
@@ -47,8 +47,6 @@ namespace Microsoft.Extensions.Azure
                 return;
             }
 
-            // The same logic is used in the Azure.Monitor.OpenTelemetry.AspNetCore.
-            // If you make any changes here, please check if they are applicable there as well.
             var logger = _loggers.GetOrAdd(eventData.EventSource.Name, name => _loggerFactory.CreateLogger(ToLoggerName(name)));
             logger.Log(MapLevel(eventData.Level), new EventId(eventData.EventId, eventData.EventName), new EventSourceEvent(eventData), null, _formatMessage);
         }
@@ -85,9 +83,6 @@ namespace Microsoft.Extensions.Azure
             }
         }
 
-        private static string FormatMessage(EventSourceEvent eventSourceEvent, Exception exception)
-        {
-            return EventSourceEventFormatting.Format(eventSourceEvent.EventData);
-        }
+        private static string FormatMessage(EventSourceEvent eventSourceEvent, Exception _) => eventSourceEvent.Format();
     }
 }
